+++ tests/libtest/first.h
@@ -197,16 +197,16 @@ void ws_close(CURL *curl);  /* just close the connection */
 
 /* ---------------------------------------------------------------- */
 
-#define exe_easy_setopt(A, B, C, Y, Z)                  \
-  do {                                                  \
-    CURLcode ec = curl_easy_setopt(A, B, C);            \
-    if(ec != CURLE_OK) {                                \
-      curl_mfprintf(stderr,                             \
-                    "%s:%d curl_easy_setopt() failed, " \
-                    "with code %d (%s)\n",              \
-                    Y, Z, ec, curl_easy_strerror(ec));  \
-      result = ec;                                      \
-    }                                                   \
+#define exe_easy_setopt(A, B, C, Y, Z)                      \
+  do {                                                      \
+    CURLcode ec = curl_easy_setopt(A, B, C);                \
+    if(ec != CURLE_OK) {                                    \
+      curl_mfprintf(stderr,                                 \
+                    "%s:%d curl_easy_setopt() failed, "     \
+                    "with code %d (%s)\n",                  \
+                    Y, Z, (int)ec, curl_easy_strerror(ec)); \
+      result = ec;                                          \
+    }                                                       \
   } while(0)
 
 #define res_easy_setopt(A, B, C) \
@@ -224,16 +224,16 @@ void ws_close(CURL *curl);  /* just close the connection */
 
 /* ---------------------------------------------------------------- */
 
-#define exe_multi_setopt(A, B, C, Y, Z)                  \
-  do {                                                   \
-    CURLMcode ec = curl_multi_setopt(A, B, C);           \
-    if(ec != CURLM_OK) {                                 \
-      curl_mfprintf(stderr,                              \
-                    "%s:%d curl_multi_setopt() failed, " \
-                    "with code %d (%s)\n",               \
-                    Y, Z, ec, curl_multi_strerror(ec));  \
-      result = TEST_ERR_MULTI;                           \
-    }                                                    \
+#define exe_multi_setopt(A, B, C, Y, Z)                      \
+  do {                                                       \
+    CURLMcode ec = curl_multi_setopt(A, B, C);               \
+    if(ec != CURLM_OK) {                                     \
+      curl_mfprintf(stderr,                                  \
+                    "%s:%d curl_multi_setopt() failed, "     \
+                    "with code %d (%s)\n",                   \
+                    Y, Z, ec, curl_multi_strerror(ec));      \
+      result = TEST_ERR_MULTI;                               \
+    }                                                        \
   } while(0)
 
 #define res_multi_setopt(A, B, C) \
@@ -536,16 +536,16 @@ void ws_close(CURL *curl);  /* just close the connection */
 
 /* ---------------------------------------------------------------- */
 
-#define exe_global_init(A, Y, Z)                        \
-  do {                                                  \
-    CURLcode ec = curl_global_init(A);                  \
-    if(ec != CURLE_OK) {                                \
-      curl_mfprintf(stderr,                             \
-                    "%s:%d curl_global_init() failed, " \
-                    "with code %d (%s)\n",              \
-                    Y, Z, ec, curl_easy_strerror(ec));  \
-      result = ec;                                      \
-    }                                                   \
+#define exe_global_init(A, Y, Z)                            \
+  do {                                                      \
+    CURLcode ec = curl_global_init(A);                      \
+    if(ec != CURLE_OK) {                                    \
+      curl_mfprintf(stderr,                                 \
+                    "%s:%d curl_global_init() failed, "     \
+                    "with code %d (%s)\n",                  \
+                    Y, Z, (int)ec, curl_easy_strerror(ec)); \
+      result = ec;                                          \
+    }                                                       \
   } while(0)
 
 #define res_global_init(A) \
